+++ src/main/java/duke/ui/Ui.java
@@ -1,4 +1,8 @@
-import java.util.ArrayList;
+package duke.ui;
+
+import duke.task.Task;
+import duke.task.TaskList;
+
 import java.util.Scanner;
 
 public class Ui {
